+++ CHANGELOG.md
@@ -3,9 +3,10 @@
 This changelog only contains the changes that are unreleased. For changes for individual releases, please visit the
 [releases](https://github.com/ATLauncher/ATLauncher/releases) page on GitHub.
 
-## 3.4.13.11
+## 3.4.14.0
 
 ### New Features
+- Add in setting for Default Export Format
 
 ### Fixes
 - Issue when adding mod from Modrinth to a Quilt loader instance not showing Fabric versions